+++ Framework/Interfaces/IGame.cs
@@ -33,10 +33,22 @@
         GraphicsDevice GraphicsDevice { get; }
 
         /// <summary>
+        /// Gets the graphics settings.
+        /// </summary>
+        /// <value>The graphics settings.</value>
+        GraphicsSettings GraphicsSettings { get; }
+
+        /// <summary>
         /// Gets a value indicating whether this game should instantiate prefabs.
         /// </summary>
         /// <value><c>true</c> if this game should instantiate prefabs; otherwise, <c>false</c>.</value>
         bool InstantiatePrefabs { get; }
+
+        /// <summary>
+        /// Gets the save data manager.
+        /// </summary>
+        /// <value>The save data manager.</value>
+        ISaveDataManager SaveDataManager { get; }
 
         /// <summary>
         /// Gets the game settings.
@@ -49,5 +61,10 @@
         /// </summary>
         /// <value>The sprite batch.</value>
         SpriteBatch SpriteBatch { get; }
+
+        /// <summary>
+        /// Saves the applies graphics settings.
+        /// </summary>
+        void SaveAndApplyGraphicsSettings();
     }
 }